requirements.txt
torch>=2.1.0
transformers>=4.45.0
trl>=0.29.0
datasets>=2.14.0
accelerate>=0.25.0
peft>=0.7.0
Pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
